% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runonce-package.R
\docType{package}
\name{runonce-package}
\alias{runonce}
\alias{runonce-package}
\title{runonce: Run Once and Save Output}
\description{
Run some long-running code once and save the result. If the result
    already exists (has already been run and saved), then just read from this
    instead of computing again.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
