% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runonce-package.R
\docType{package}
\name{runonce-package}
\alias{runonce}
\alias{runonce-package}
\title{runonce: Run Once and Save Result}
\arguments{
\item{timing}{Whether to print timing of running code? Default is \code{TRUE}.}
}
\description{
Package 'runonce' helps automating the saving of long-running code
    to help running the same code multiple times. If you run some long-running 
    code once, it saves the result in a file on disk. Then, if the result 
    already exists, i.e. if the code has already been run and its output has 
    already been saved, it just reads the result from the stored file instead 
    of running the code again.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/runonce}
  \item Report bugs at \url{https://github.com/privefl/runonce/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
